Avoid rounding error in circular arc
It's possible the first point of the bezier on a circular arc won't be the specified first point, which will cause problems with the new system where all segments contain all their points (so red anchors are just two segments overlaid perfectly). This fixes this potential problem by explicitly setting the first point, like we were already doing with the last point.